move:#24 파일명 변경 및 구조 임시 정리
