+++ src/constants/query-keys.ts
@@ -1,3 +1,4 @@
 export const queryKeys = {
-  MY_IN_PROGRESS_CHALLENGE: 'my-in-progress-challenge'
+  MY_IN_PROGRESS_CHALLENGE: 'my-in-progress-challenge',
+  MY_COMPLETED_CHALLENGE: 'my-completed-challenge'
 };
